moved working code from /api to api wrapper, to allow other files to use the wrapper
